ai-sentiment-analyzer/requirements.txt
flask
transformers
torch
pymongo
python-dotenv
gunicorn
